Update and fix WaitForReply*, API modified
Complete changeover to replyLock, fixing panic

Replace 'ReplyInProgress' with 'ConversationInterrupted' return
values for the WaitForReply* functions.. This is better
semantics for WaitForReply, and allows longer timeouts in
cases where the user can optionally continue with follow-ups on
a command in progress. If the user issues a new, unrelated command,
the method returns 'ConversationInterrupted', and the plugin can
just exit silently if no reply was required. Eventually a new
WaitForReplyMulti may be implemented, for cases where the user
could issue any number of follow-on commands, perhaps in query
situations where output is limited to XX entries, and the user
wishes to further constrain the output.