Experimental "it" support, linting cleanups
Added short-term memories support to the brain; a simple
hash with no backing store. Short-term memories expire after
a fixed amount of time. Added functions "Remember", "RememberIt",
and "Recall". Implemented to support "it".

Modified dispatch to check if the user mentioned "it" in their last
message, in which case the robot checks to see if "it" is in it's
short term memories; if so, it does a string replace and checks
the message against commands. Note that "it" is currently only
expanded and checked against commands, and not ambient messages
or replies being waited for.

Updated the "lists" plugin to RememberIt(...)

Various renames for linting