Experimental support for "it"
Add optional 'Nouns []string' for CommandMatchers. If set, the
robot will keep references to types of nouns in short-term memory;
e.g. if the first argument to a command is a server, and Nouns[0]
is "server", then the robot will store whatever matches the first
capture group as the "server" noun. So, if the user provided
"web1.my.dom" for the first argument, the user can then use "it"
in the context of a server in another command, and the robot will
substitute "web1.my.dom".

Add "Nouns" declaration for included lists and rubydemo plugins.
The robot will keep track of which list and which item are
being discussed, and the item argument to "remember" (from rubydemo)
can be used interchangably with items being added/removed with lists.

Updated script libraries with RememberNoun convenience function.
Names of nouns in short term memories are namespaced with a "noun:"
prefix, which the RememberNoun method adds.