Updates to experimental contextual memory support
Rename capture group labels to "Contexts"; if a capture group
with a Context label corresponds to a blank argument, or "it", the
robot will use it's last contextual memory to fill in the gap. This
changes the behavior of the "lists" plugin; now the last list should
be referenced by "the list" instead of "it"; e.g. when you say
"Add milk to the grocery list", you can follow up with "Add bananas
to the list", but not "Add bananas to it" (because the word "list"
is required)