Merge callplugin implementations
It's a little gross, but callplugin now just has a couple of
conditional statements that changes the calling convention based
on platform. On Unix, the kernel calls the interpreter, but on
Windows the interpreter is executed with the script as first
argument. When running powershell plugins on Linux (I'm not
kidding) the arguments still needed to be massaged, so
getInterpreter and fixInterpreterArgs were both needed. This
greatly reduces code duplication at the expense of an only
slightly ugly if statement.