Updates for hearing bots, plugin locking fix
Previously, if an app integratin bot sent a message to a channel,
the bot couldn't hear it at all. Now the bot gets a message from
a generated User name. Note that for the robot to 'see' a newly
added app, it'll need to be restarted.

The plugins data structure was being protected with the robot lock,
unnecessarily. Renamed plugins to pluginlist with embedded mutex.
Also, we mostly just copy the current plugin list under lock, so
we hold the lock for shorter durations.