Fix race condition in waitreply
It was possible (though highly improbable) that a timeout would
expire just as a reply came in. If dispatch got the replies lock
first, it would try to send a reply on a channel that's not being
listened on, because timeout was just deleting the hash entry
without first checking that it's still there in the first place.
Changed timeout behavior to first check that the hash entry exists,
and if not it listens on the channel for the reply.
Thanks to Dr. Marmorstein and his students at Longwood who asked
me about Go, which led me to describe this algorithm and later
realize it had a race condition.