Do nothing on multiple command matches
There were two basic issues:
- checkPluginMatchers was frequently returning in the middle of
  the plugin loop, which was ugly (it can still do this in a few
  cases)
- If two plugins both match a command, they run in sequence; this
  could allow a potentially malicious plugin to interact with the
  user and potentially obtain sensitive information when the user
  believes they're interacting with a different plugin

I believe that a "secure" Gopherbot installation should never mix
trusted and untrusted plugins - but I still try to take that case
in to account. In any event, I consider multiple commands matching
an error, so we emit an error message to the user and log an error.