Re-jigger plugin locking, start work on CallPlugin
Plugin data structure locking was a total mess. With this patch,
all references to a plugin or the plugin list should return a
pointer or a copy of an array of pointers. Functions will run with
whatever plugin / list was valid when obtained, and if the plugin
or list is replaced the old plugin structs will be garbage
collected after the function runs.

Started work on CallPlugin for external plugins, left a TODO in
http.go