Add message delay loop
When nlopes/slack switched to gorilla websocket, it highlighted
a bug in Gopherbot where the robot was sending messages too fast
and being rate limited by Slack. Previously the library was
transparently reconnecting and sending the message, so the bug
wasn't apparent. This patch creates a sendMessages infinite loop
goroutine that insures a 1 second delay after every message, to
avoid hitting the rate limit.