Rework base64 in libraries
Not sure why base64 encoding and decoding were so heavily used
previously, but every language but Bash should be able to handle
encoding / decoding JSON without needing to resort to base64 for
strings. The Bash library now uses base64 for encoding all strings
sent to the robot, but flags it with a Base64 boolean instead of
the "base64:" string prefix. This should make all code simpler
and more correct.