Authorizer/Elevator changes, getPluginByName
After much pondering, it doesn't make any sense to check visibility
of the authorizer or elevator plugins in the channel where
authorization / elevation is being requested, so this check was
removed. This means an administrator can restrict where e.g.
configuration commands are visible (configure duo is an example)
without restricting the plugins that can use them. Just a think-o
that needed to be fixed.

Fix some spots in the code that were doing manual plugin lookups
to use currentPlugins.getPluginBy(Name|ID) instead.