Properly implement Variable & Raw
For Gopherbot to be more uniform across connectors, the characters
"_", "*", and "`" should pass through unharmed to the user, as
those characters can be perfectly legitimate in a variety of
contexts. This commit 'escapes' those characters so they pass
through for Variable, but leaves them alone when Format == Raw,
which preserves the old behavior. This is a minor breaking change
for the benefit of correctness; plugin authors can use the 'raw'
format to restore the old behavior. Before merging, a convenience
function should be provided for getting a Raw bot. Note, however,
that a well-behaved plugin will check the value of Protocol before
sending a raw message, so it can use appropriate characters.