Message formatting updates and more (#35)
* Properly implement Variable & Raw

For Gopherbot to be more uniform across connectors, the characters
"_", "*", and "`" should pass through unharmed to the user, as
those characters can be perfectly legitimate in a variety of
contexts. This commit 'escapes' those characters so they pass
through for Variable, but leaves them alone when Format == Raw,
which preserves the old behavior. This is a minor breaking change
for the benefit of correctness; plugin authors can use the 'raw'
format to restore the old behavior. Before merging, a convenience
function should be provided for getting a Raw bot. Note, however,
that a well-behaved plugin will check the value of Protocol before
sending a raw message, so it can use appropriate characters.

* New MessageFormat method, doc update

* Implement DefaultMessageFormat

Python, Ruby and PowerShell still need to be tested.

The robot defaults to sending 'Raw' text to the protocol, but
can be configured to send 'Variable' or 'Fixed' messages by default.

* Message format fixes

* The long-awaited CHUCK NORRIS plugin

* Move testing & sample plugins to samples

* Formatting, case sensitivity, debugging

Added message formatting to test & terminal connectors:
- FIXED (-> all upper)
- variable (-> all lower)
- Raw (mixed case)

Updated tests & added formatting tests

Force lowercase for attributes in for Get*Attribute, for
case-insensitivity.

No longer 'reload' in debug plugin, it didn't help anything.
Tweak regex.

Reorganize testing plugins into 'samples'.

Added 'whoami' to the ping plugin.

* Update Go version for build

* Fix PowerShell lib, add format.ps1