MessageHeard, use PostMessage
For some unknown reason, there was an issue with the robot saying
something that never reached the channel. Since there's no error
checking on the old RTM.SendMessage, we switch to API.PostMessage
and check for errors, with exponential backoff and retry.

At times a plugin will run but take a while to actually Say or
Reply anything, with no feedback to the user. The Slack connector
was formerly sending a typing notifier with each outgoing message,
but the delay was so short it was frequently not seen. Now,
MessageHeard is part of the Connector interface, and is triggered
whenever a command or message matches, before the plugin is run;
this way the user is given immediate feedback, even if the plugin
takes several seconds to generate output. For the terminal and
test connectors it's a no-op.