Add FinalTask(...), update pipeline algorithm
Library method FinalTask lets tasks add other tasks to a list
of cleanup tasks that always run.

When an added task adds multiple tasks in the middle of a pipeline,
those tasks are executed as a separate, sub-pipeline. If the
current task is the last task in a pipeline, the new tasks are
just appended.