Re-fix old spacing issue nasty hack
About a year ago there was a user issue where copying and pasting
a command caused a bogus failure to match because most regexes
have single spaces between words. At the time the very hacky fix
was 'massaging' regexes so that single space matches would match
one or more spaces - yuck. I always felt the 'massageRegexp'
function was fragile, and for some reason today it occurred to
me - for messages from a user, we should just replace multiple
spaces with a single space. Easy enough, and not fragile - BUT,
that means we have to reject regexes with multiple spaces.