Update use of environment vars
A few secrets need to be configured or provided in the environment;
others can and should be stored in an encrypted brain. For those
that need to be provided, the most secure method is providing them
as environment vars in the systemd unit file only readable by root;
this patch clears those environment vars after use, ensuring that
child processes spawned after the robot starts are unable to read
them.