Privilege separation
When the gopherbot binary is installed suid to another user, that
user is considered the privileged user that can read the ".env"
file and read/write the contents of the brain, and do any other
file operations in Go. Before calling external scripts, the robot
will spawn a new thread and drop privileges to the original uid.