Massive locking update
Previously the current configuration could change during the
running of a pipeline, and all accesses to the current config
had to be taken under lock. Now handler.go puts a pointer to the
current config in the botContext (retrieved under read lock),
and all locking has been removed elsewhere, since nothing will
change in the configuration pointed to by the context.